# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Cursos de Finanzas",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/progresos": {
            "get": {
                "tags": [
                    "Progresos"
                ],
                "summary": "Listar progresos con filtros opcionales",
                "operationId": "ff62fe15e47bd25e3867549049148d73",
                "parameters": [
                    {
                        "name": "id_inscripcion",
                        "in": "query",
                        "description": "Filtrar por ID de inscripción",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_modulo",
                        "in": "query",
                        "description": "Filtrar por ID de módulo",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "completado",
                        "in": "query",
                        "description": "Filtrar por completado (true/false)",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "activo",
                        "in": "query",
                        "description": "Filtrar por estado activo/inactivo",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad por página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de progresos"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Progresos"
                ],
                "summary": "Registrar un nuevo progreso",
                "operationId": "7fb276416a12bab22de33f907c14b356",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_inscripcion",
                                    "id_modulo",
                                    "fecha_inicio"
                                ],
                                "properties": {
                                    "id_inscripcion": {
                                        "type": "integer"
                                    },
                                    "id_modulo": {
                                        "type": "integer"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "calificacion": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tiempo_total": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Progreso creado exitosamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/progresos/{id}": {
            "put": {
                "tags": [
                    "Progresos"
                ],
                "summary": "Actualizar un progreso existente",
                "operationId": "f09d5c89b2b41233f9228e9c3eeb3bf4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del progreso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_inscripcion": {
                                        "type": "integer"
                                    },
                                    "id_modulo": {
                                        "type": "integer"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "completado": {
                                        "type": "boolean"
                                    },
                                    "calificacion": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tiempo_total": {
                                        "type": "integer"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Progreso actualizado exitosamente"
                    },
                    "404": {
                        "description": "Progreso no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Progresos"
                ],
                "summary": "Eliminar un progreso",
                "operationId": "209296e899f2fecb1943106d59f7fc4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del progreso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progreso eliminado correctamente"
                    },
                    "404": {
                        "description": "Progreso no encontrado"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token JWT aquí. Ejemplo: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}